requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pinecone-client>=3.0.0
pypdf>=3.0.0
